pyproject.toml
[tool.poetry]
name = "sead-compare"
version = "0.1.3"
description = ""
authors = ["Roger Mähler <roger.mahler@hotmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
data-diff = "^0.8.4"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
click = "^8.1.7"
tqdm = "^4.66.1"
jinja2 = "^3.1.2"

[tool.poetry.scripts]
sead-differ = "src.sead_data_diff.src.:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pylint = "^2.17.6"
isort = "^5.12.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
